+++ apps/website/src/components/demo/examples/default-card.tsx
@@ -1,0 +1,11 @@
+import { Card } from '@vapor-ui/core';
+
+export default function DefaultCard() {
+    return (
+        <Card.Root>
+            <Card.Header>Basic Template</Card.Header>
+            <Card.Body>This is a Basic Template</Card.Body>
+            <Card.Footer style={{ display: 'flex', justifyContent: 'center' }}>Footer</Card.Footer>
+        </Card.Root>
+    );
+}