+++ apps/website/src/components/demo/examples/default-text-input.tsx
@@ -1,0 +1,10 @@
+import { TextInput } from '@vapor-ui/core';
+
+export default function DefaultTextInput() {
+    return (
+        <TextInput.Root>
+            <TextInput.Label>Label</TextInput.Label>
+            <TextInput.Field />
+        </TextInput.Root>
+    );
+}