+++ apps/website/src/components/foundation-size-tabs/size-cell.tsx
@@ -22,10 +22,10 @@ const SemanticSizeRow = ({ sizeType = 'width', name, value }: SemanticSizeRowPro
             <td>
                 <div
                     style={{
-                        width: isBorderRadius ? ' 5rem' : `var(--${name})`,
-                        height: isBorderRadius ? ' 5rem' : `var(--${name})`,
-                        backgroundColor: 'var(--blue-100)',
-                        borderRadius: isBorderRadius ? `var(--${name})` : '0',
+                        width: isBorderRadius ? ' 5rem' : `var(${name})`,
+                        height: isBorderRadius ? ' 5rem' : `var(${name})`,
+                        backgroundColor: 'var(--vapor-color-blue-100)',
+                        borderRadius: isBorderRadius ? `var(${name})` : '0',
                     }}
                 />
             </td>