+++ packages/core/src/components/avatar/avatar.css.ts
@@ -1,13 +1,14 @@
 import { createVar } from '@vanilla-extract/css';
+import type { RecipeVariants } from '@vanilla-extract/recipes';
 import { recipe } from '@vanilla-extract/recipes';
 
-import { vars } from '~/styles/contract.css';
-import { layerStyle } from '~/styles/layers.css';
+import { layerStyle } from '~/styles/utils/layer-style.css';
+import { vars } from '~/styles/vars.css';
 
 const radii = createVar('avatar-border-radius');
 
 export const root = recipe({
-    base: layerStyle('component', {
+    base: layerStyle('vapor-component', {
         display: 'inline-flex',
         flexShrink: 0,
         alignItems: 'center',
@@ -16,35 +17,36 @@ export const root = recipe({
         border: `1px solid ${vars.color.border.normal}`,
         padding: vars.size.space['000'],
         overflow: 'hidden',
+        verticalAlign: 'top',
     }),
 
     defaultVariants: { size: 'md', shape: 'square' },
     variants: {
         size: {
-            sm: layerStyle('component', {
+            sm: layerStyle('vapor-component', {
                 width: vars.size.dimension[300],
                 height: vars.size.dimension[300],
                 vars: { [radii]: vars.size.borderRadius[200] },
             }),
-            md: layerStyle('component', {
+            md: layerStyle('vapor-component', {
                 width: vars.size.dimension[400],
                 height: vars.size.dimension[400],
                 vars: { [radii]: vars.size.borderRadius[300] },
             }),
-            lg: layerStyle('component', {
+            lg: layerStyle('vapor-component', {
                 width: vars.size.dimension[500],
                 height: vars.size.dimension[500],
                 vars: { [radii]: vars.size.borderRadius[400] },
             }),
-            xl: layerStyle('component', {
+            xl: layerStyle('vapor-component', {
                 width: vars.size.dimension[600],
                 height: vars.size.dimension[600],
                 vars: { [radii]: vars.size.borderRadius[400] },
             }),
         },
         shape: {
-            square: layerStyle('component', { borderRadius: radii }),
-            circle: layerStyle('component', { borderRadius: '50%' }),
+            square: layerStyle('vapor-component', { borderRadius: radii }),
+            circle: layerStyle('vapor-component', { borderRadius: '50%' }),
         },
     },
 });
@@ -52,7 +54,7 @@ export const root = recipe({
 export const fallbackBgVar = createVar('fallback-background-color');
 
 export const fallback = recipe({
-    base: layerStyle('component', {
+    base: layerStyle('vapor-component', {
         display: 'inline-flex',
         alignItems: 'center',
         justifyContent: 'center',
@@ -66,25 +68,25 @@ export const fallback = recipe({
     defaultVariants: { size: 'md' },
     variants: {
         size: {
-            sm: layerStyle('component', {
+            sm: layerStyle('vapor-component', {
                 fontSize: vars.typography.fontSize['050'],
                 lineHeight: vars.typography.lineHeight['050'],
                 fontWeight: vars.typography.fontWeight['500'],
                 letterSpacing: vars.typography.letterSpacing['000'],
             }),
-            md: layerStyle('component', {
+            md: layerStyle('vapor-component', {
                 fontSize: vars.typography.fontSize['075'],
                 lineHeight: vars.typography.lineHeight['075'],
                 fontWeight: vars.typography.fontWeight['500'],
                 letterSpacing: vars.typography.letterSpacing['100'],
             }),
-            lg: layerStyle('component', {
+            lg: layerStyle('vapor-component', {
                 fontSize: vars.typography.fontSize['200'],
                 lineHeight: vars.typography.lineHeight['200'],
                 fontWeight: vars.typography.fontWeight['700'],
                 letterSpacing: vars.typography.letterSpacing['100'],
             }),
-            xl: layerStyle('component', {
+            xl: layerStyle('vapor-component', {
                 fontSize: vars.typography.fontSize['300'],
                 lineHeight: vars.typography.lineHeight['300'],
                 fontWeight: vars.typography.fontWeight['700'],
@@ -94,9 +96,12 @@ export const fallback = recipe({
     },
 });
 
-export const image = layerStyle('component', {
+export const image = layerStyle('vapor-component', {
     display: 'inline',
     objectFit: 'cover',
     width: '100%',
     height: '100%',
 });
+
+export type RootVariants = NonNullable<RecipeVariants<typeof root>>;
+export type FallbackVariants = NonNullable<RecipeVariants<typeof fallback>>;