+++ packages/core/src/components/button/button.tsx
@@ -1,17 +1,17 @@
+'use client';
+
 import type { ComponentPropsWithoutRef } from 'react';
 import { forwardRef } from 'react';
 
+import { Primitive } from '@radix-ui/react-primitive';
 import clsx from 'clsx';
 
-import { vapor } from '~/libs/factory';
-import type { MergeRecipeVariants } from '~/libs/recipe';
 import { createSplitProps } from '~/utils/create-split-props';
 
+import type { ButtonVariants } from './button.css';
 import * as styles from './button.css';
 
-type ButtonPrimitiveProps = Omit<ComponentPropsWithoutRef<typeof vapor.button>, 'color'>;
-type ButtonVariants = MergeRecipeVariants<typeof styles.root>;
-
+type ButtonPrimitiveProps = Omit<ComponentPropsWithoutRef<typeof Primitive.button>, 'color'>;
 interface ButtonProps extends ButtonPrimitiveProps, ButtonVariants {}
 
 const Button = forwardRef<HTMLButtonElement, ButtonProps>(
@@ -24,14 +24,14 @@ const Button = forwardRef<HTMLButtonElement, ButtonProps>(
         ]);
 
         return (
-            <vapor.button
+            <Primitive.button
                 ref={ref}
                 className={clsx(styles.root(variantsProps), className)}
                 data-disabled={otherProps.disabled}
                 {...otherProps}
             >
                 {children}
-            </vapor.button>
+            </Primitive.button>
         );
     },
 );