+++ packages/core/src/components/grid/grid.css.ts
@@ -1,13 +1,14 @@
 import { createVar } from '@vanilla-extract/css';
+import type { RecipeVariants } from '@vanilla-extract/recipes';
 import { recipe } from '@vanilla-extract/recipes';
 
-import { layerStyle } from '~/styles/layers.css';
+import { layerStyle } from '~/styles/utils/layer-style.css';
 
 export const gridTemplateRows = createVar({ inherits: false, syntax: '*' }, 'template-rows');
 export const gridTemplateColumns = createVar({ inherits: false, syntax: '*' }, 'template-columns');
 
 export const root = recipe({
-    base: layerStyle('component', {
+    base: layerStyle('vapor-component', {
         gridTemplateRows: gridTemplateRows,
         gridTemplateColumns: gridTemplateColumns,
     }),
@@ -15,10 +16,10 @@ export const root = recipe({
     defaultVariants: { flow: 'row' },
     variants: {
         flow: {
-            row: layerStyle('component', { gridAutoFlow: 'row' }),
-            column: layerStyle('component', { gridAutoFlow: 'column' }),
-            'row-dense': layerStyle('component', { gridAutoFlow: 'row dense' }),
-            'column-dense': layerStyle('component', { gridAutoFlow: 'column dense' }),
+            row: layerStyle('vapor-component', { gridAutoFlow: 'row' }),
+            column: layerStyle('vapor-component', { gridAutoFlow: 'column' }),
+            'row-dense': layerStyle('vapor-component', { gridAutoFlow: 'row dense' }),
+            'column-dense': layerStyle('vapor-component', { gridAutoFlow: 'column dense' }),
         },
     },
 });
@@ -26,7 +27,9 @@ export const root = recipe({
 export const gridItemRowSpan = createVar({ inherits: false, syntax: '*' }, 'item-row-span');
 export const gridItemColSpan = createVar({ inherits: false, syntax: '*' }, 'item-col-span');
 
-export const item = layerStyle('component', {
+export const item = layerStyle('vapor-component', {
     gridRow: gridItemRowSpan,
     gridColumn: gridItemColSpan,
 });
+
+export type RootVariants = NonNullable<RecipeVariants<typeof root>>;