+++ packages/core/src/components/text/text.tsx
@@ -1,30 +1,35 @@
+'use client';
+
+import type { ComponentPropsWithoutRef } from 'react';
 import { forwardRef } from 'react';
 
+import { Primitive } from '@radix-ui/react-primitive';
 import clsx from 'clsx';
 
-import { vapor } from '~/libs/factory';
-import { foregroundSprinkles, typographySprinkles } from '~/styles/sprinkles';
-import type { Foreground, Typography } from '~/styles/sprinkles';
+import type { Foregrounds } from '~/styles/mixins/foreground.css';
+import { foregrounds } from '~/styles/mixins/foreground.css';
+import { type Typography, typography } from '~/styles/mixins/typography.css';
 
-interface TextProps extends React.ComponentPropsWithoutRef<typeof vapor.span> {
-    typography?: Typography;
-    foreground?: Foreground;
+type TextPrimitiveProps = ComponentPropsWithoutRef<typeof Primitive.span>;
+interface TextProps extends TextPrimitiveProps {
+    foreground?: Foregrounds['color'];
+    typography?: Typography['style'];
 }
 
 const Text = forwardRef<HTMLSpanElement, TextProps>(
-    ({ typography, foreground, children, className, ...props }, ref) => {
+    ({ typography: typographyStyle, foreground, className, children, ...props }, ref) => {
         return (
-            <vapor.span
+            <Primitive.span
                 ref={ref}
                 className={clsx(
+                    typography({ style: typographyStyle }),
+                    foregrounds({ color: foreground }),
                     className,
-                    typographySprinkles({ typography }),
-                    foregroundSprinkles({ foreground }),
                 )}
                 {...props}
             >
                 {children}
-            </vapor.span>
+            </Primitive.span>
         );
     },
 );