+++ packages/core/src/styles/layers.css.ts
@@ -1,5 +1,11 @@
-import type { StyleRule } from '@vanilla-extract/css';
-import { globalLayer, style } from '@vanilla-extract/css';
+import { globalLayer } from '@vanilla-extract/css';
+
+export const layerName = {
+    theme: 'vapor-theme',
+    reset: 'vapor-reset',
+    component: 'vapor-component',
+    utilities: 'vapor-utilities',
+} as const;
 
 /**
  * The order of the declaration means the priority of the layer.
@@ -17,25 +23,10 @@ import { globalLayer, style } from '@vanilla-extract/css';
  * @layer vapor-theme, theme, vapor-reset, base, vapor-component, components, vapor-utilities, utilities;
  *
  */
-const theme = globalLayer('vapor-theme');
-const reset = globalLayer('vapor-reset');
-const component = globalLayer('vapor-component');
-const utilities = globalLayer('vapor-utilities');
 
-const layerMap = { reset, theme, component, utilities };
-
-export const layerStyle = (
-    layer: 'theme' | 'reset' | 'component' | 'utilities',
-    rule: StyleRule,
-    debugId?: string,
-) =>
-    style(
-        {
-            '@layer': {
-                [layerMap[layer]]: rule,
-            },
-        },
-        debugId,
-    );
+const theme = globalLayer(layerName.theme);
+const reset = globalLayer(layerName.reset);
+const component = globalLayer(layerName.component);
+const utilities = globalLayer(layerName.utilities);
 
 export const layers = { theme, reset, component, utilities };